+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added participant count field
 * Fixed custom link field so that the raw value also contains the html link (#55).
 * Fix for Field filter. It did not work on a custom field that has the type country.
+* Added image field output handler (!61)
 
 # Version 1.16